aider: Set the authentication state to true if the email is in the whitelist, otherwise show an error message.
# Aider chat conversation:

USER: how should be the implementation of commented line in auth.tsx if (data.length > 0) {
      // The email is in the whitelist.
      // Here you can set the authentication state.

    } else {
      // The email is not in the whitelist.
      // Show an error message.
    }

ASSISTANT: To implement the commented lines, we need to define what the authentication state is and how to set it. For simplicity, let's assume we have a function `setAuthenticated` that takes a boolean to set the authentication state.

For the error message, we can use the `alert` function to show a simple popup with the error message.

Here are the changes:

```tsx
pages/auth.tsx
<<<<<<< ORIGINAL
    if (data.length > 0) {
      // The email is in the whitelist.
      // Here you can set the authentication state.
    } else {
      // The email is not in the whitelist.
      // Show an error message.
    }
=======
    if (data.length > 0) {
      // The email is in the whitelist.
      setAuthenticated(true);
    } else {
      // The email is not in the whitelist.
      alert("The email is not in the whitelist.");
    }
>>>>>>> UPDATED
```

Please note that `setAuthenticated` is a hypothetical function. You need to replace it with the actual function or method that sets the authentication state in your application.